requirements.txt
pip~=23.2.1
wheel~=0.41.2
cryptography~=41.0.5
Pillow~=8.2.0
Django~=3.1.7
django-daraja==1.3.0

Jinja2~=3.1.2
requests~=2.31.0
djangorestframework~=3.14.0